Nesting Custom Type-Using @OneToOne(cascade=CascadeType.ALL)
In hibernate there are 3 ways to create one-to-one relationships between two entities. Either way we have to use @OneToOne annotation.

First technique is most widely used and uses a foreign key column in one of the tables.
Second technique uses a rather known solution of having a third table to store mapping between first two tables.
Third technique is something new which uses a common primary key value in both the tables.

1. Hibernate one to one mapping with foreign key association
In this kind of association, a foreign key column is created in owner entity. For example, if we make EmployeeEntity owner, then a extra column "ACCOUNT_ID" will be created in Employee table. This column will store the foreign key for Account table.